# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Number Book Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books/{bookId}/entries:
    get:
      operationId: listEntries
      tags: [entries]
      summary: List all entries for a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
  /books/{bookId}/entries/{entryId}/regenerate-title:
    post:
      operationId: regenerateEntryTitle
      tags: [entries]
      summary: Regenerate the title for a single entry using AI
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry with regenerated title
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
  /books/{bookId}/entries/{entryId}/generate:
    post:
      operationId: generateEntry
      tags: [entries]
      summary: Generate content for a single entry
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Entry with generated content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
  /books/{bookId}/entries/{entryId}:
    patch:
      operationId: updateEntry
      tags: [entries]
      summary: Update an entry (lock/unlock, edit content)
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntryUpdate"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
components:
  schemas:
...
    Entry:
      type: object
      required: [id, bookId, position, title, status, isLocked, createdAt, updatedAt]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        position:
          type: integer
        title:
          type: string
        content:
          type: ["string", "null"]
        contentJson:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, generating, done, failed]
        isLocked:
          type: boolean
        wordCount:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    EntryUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        contentJson:
          type: string
        isLocked:
          type: boolean
        status:
          type: string
...